Reorganize cmake project and include scoping
