+++ CMakeLists.txt
@@ -11,7 +11,7 @@ project(HumanDynamicsEstimation
 # PROJECT CONFIGURATION
 # =====================
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 if(BUILD_DOCS)
