+++ HumanDynamicsEstimationLibrary/include/hde/Utils.hpp
