+++ msgs/yarp/thrift/CMakeLists.txt
@@ -3,25 +3,28 @@
 # GNU Lesser General Public License v2.1 or any later version.
 
 set(LIBRARY_TARGET_NAME HumanStateMsg)
-set (HDE_PROJECT_NAME HumanDynamicsEstimation)
 
 
 yarp_idl_to_dir(HumanDynamicsEstimation/HumanState.thrift
-  ${CMAKE_CURRENT_BINARY_DIR}/autogenerated
+  OUTPUT_DIR ${CMAKE_CURRENT_BINARY_DIR}/autogenerated
   SOURCES_VAR ${LIBRARY_TARGET_NAME}_GEN_SRC
   HEADERS_VAR ${LIBRARY_TARGET_NAME}_GEN_HDR
-  INCLUDE_DIRS_VAR ${LIBRARY_TARGET_NAME}_INCLUDE_DIRS)
+  INCLUDE_DIRS_VAR ${LIBRARY_TARGET_NAME}_INCLUDE_DIRS
+  CMAKE_SCRIPTS_VAR ${LIBRARY_TARGET_NAME}_CMAKE_SCRIPTS)
+foreach(_script ${${LIBRARY_TARGET_NAME}_CMAKE_SCRIPTS})
+  include("${_script}")
+endforeach()
 
-include_directories(${LIBRARY_TARGET_NAME}_INCLUDE_DIRS)
+# include_directories(${LIBRARY_TARGET_NAME}_INCLUDE_DIRS)
 add_library(${LIBRARY_TARGET_NAME} SHARED ${${LIBRARY_TARGET_NAME}_GEN_SRC} ${${LIBRARY_TARGET_NAME}_GEN_HDR})
-target_link_libraries(${LIBRARY_TARGET_NAME} YARP::YARP_OS)
+target_link_libraries(${LIBRARY_TARGET_NAME} YARP::YARP_OS YARP::YARP_init)
 
 set_target_properties(${LIBRARY_TARGET_NAME} PROPERTIES
                       PUBLIC_HEADER "${${LIBRARY_TARGET_NAME}_GEN_HDR}")
 
 target_include_directories(${LIBRARY_TARGET_NAME} PUBLIC
                            "$<BUILD_INTERFACE:${${LIBRARY_TARGET_NAME}_INCLUDE_DIRS}>"
-                           "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/${HDE_PROJECT_NAME}>")
+                           "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/hde/msgs>")
 
 set_property(TARGET ${LIBRARY_TARGET_NAME} PROPERTY PUBLIC_HEADER ${${LIBRARY_TARGET_NAME}_GEN_HDR})
 
@@ -31,9 +34,9 @@ install(TARGETS ${LIBRARY_TARGET_NAME}
         EXPORT ${PROJECT_NAME}
         RUNTIME DESTINATION "${CMAKE_INSTALL_BINDIR}"
         LIBRARY DESTINATION "${CMAKE_INSTALL_LIBDIR}"
-        PUBLIC_HEADER DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/${HDE_PROJECT_NAME}")
+        PUBLIC_HEADER DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/hde/msgs")
 
-set_property(GLOBAL APPEND PROPERTY ${PROJECT_NAME}_TARGETS ${LIBRARY_TARGET_NAME})
+# set_property(GLOBAL APPEND PROPERTY ${PROJECT_NAME}_TARGETS ${LIBRARY_TARGET_NAME})
 
 
 
@@ -44,12 +47,16 @@ set(LIBRARY_TARGET_NAME HumanDynamicsMsg)
 
 
 yarp_idl_to_dir(HumanDynamicsEstimation/HumanDynamics.thrift
-  ${CMAKE_CURRENT_BINARY_DIR}/autogenerated
+  OUTPUT_DIR ${CMAKE_CURRENT_BINARY_DIR}/autogenerated
   SOURCES_VAR ${LIBRARY_TARGET_NAME}_GEN_SRC
   HEADERS_VAR ${LIBRARY_TARGET_NAME}_GEN_HDR
-  INCLUDE_DIRS_VAR ${LIBRARY_TARGET_NAME}_INCLUDE_DIRS)
+  INCLUDE_DIRS_VAR ${LIBRARY_TARGET_NAME}_INCLUDE_DIRS
+  CMAKE_SCRIPTS_VAR ${LIBRARY_TARGET_NAME}_CMAKE_SCRIPTS)
+foreach(_script ${YARP_dev_idl_CMAKE_SCRIPTS})
+  include("${_script}")
+endforeach()
 
-include_directories(${LIBRARY_TARGET_NAME}_INCLUDE_DIRS)
+# include_directories(${LIBRARY_TARGET_NAME}_INCLUDE_DIRS)
 add_library(${LIBRARY_TARGET_NAME} SHARED ${${LIBRARY_TARGET_NAME}_GEN_SRC} ${${LIBRARY_TARGET_NAME}_GEN_HDR})
 target_link_libraries(${LIBRARY_TARGET_NAME} YARP::YARP_OS)
 
@@ -58,7 +65,7 @@ set_target_properties(${LIBRARY_TARGET_NAME} PROPERTIES
 
 target_include_directories(${LIBRARY_TARGET_NAME} PUBLIC
                            "$<BUILD_INTERFACE:${${LIBRARY_TARGET_NAME}_INCLUDE_DIRS}>"
-                           "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/${HDE_PROJECT_NAME}>")
+                           "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/hde/msgs>")
 
 set_property(TARGET ${LIBRARY_TARGET_NAME} PROPERTY PUBLIC_HEADER ${${LIBRARY_TARGET_NAME}_GEN_HDR})
 
@@ -68,6 +75,6 @@ install(TARGETS ${LIBRARY_TARGET_NAME}
         EXPORT ${PROJECT_NAME}
         RUNTIME DESTINATION "${CMAKE_INSTALL_BINDIR}"
         LIBRARY DESTINATION "${CMAKE_INSTALL_LIBDIR}"
-        PUBLIC_HEADER DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/${HDE_PROJECT_NAME}")
+        PUBLIC_HEADER DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/hde/msgs")
 
-set_property(GLOBAL APPEND PROPERTY ${PROJECT_NAME}_TARGETS ${LIBRARY_TARGET_NAME})
+# set_property(GLOBAL APPEND PROPERTY ${PROJECT_NAME}_TARGETS ${LIBRARY_TARGET_NAME})
