+++ wrappers/HumanDynamicsWrapper/HumanDynamicsWrapper.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "HumanDynamicsWrapper.h"
-#include <IHumanDynamics.h>
+#include <hde/interfaces/IHumanDynamics.h>
 #include <HumanDynamicsEstimation/HumanDynamics.h>
 
 #include <yarp/os/BufferedPort.h>
@@ -27,7 +27,7 @@ class HumanDynamicsWrapper::impl
 public:
     mutable std::mutex mutex;
     hde::interfaces::IHumanDynamics* humanDynamics = nullptr;
-    yarp::os::BufferedPort<human::HumanDynamics> outputPort;
+    yarp::os::BufferedPort<hde::msgs::HumanDynamics> outputPort;
 };
 
 HumanDynamicsWrapper::HumanDynamicsWrapper()
@@ -90,7 +90,7 @@ void HumanDynamicsWrapper::run()
     std::vector<std::string> jointNames = pImpl->humanDynamics->getJointNames();
 
     // Prepare the message
-    human::HumanDynamics& humanDynamicsData = pImpl->outputPort.prepare();
+    hde::msgs::HumanDynamics& humanDynamicsData = pImpl->outputPort.prepare();
 
     // Convert the joint names
     humanDynamicsData.jointNames.resize(jointTorques.size());
