+++ wrappers/HumanStateWrapper/HumanStateWrapper.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "HumanStateWrapper.h"
-#include <IHumanState.h>
+#include <hde/interfaces/IHumanState.h>
 #include <HumanDynamicsEstimation/HumanState.h>
 
 #include <yarp/os/BufferedPort.h>
@@ -24,7 +24,7 @@ class HumanStateWrapper::impl
 {
 public:
     hde::interfaces::IHumanState* iHumanState = nullptr;
-    yarp::os::BufferedPort<human::HumanState> outputPort;
+    yarp::os::BufferedPort<hde::msgs::HumanState> outputPort;
 
     // buffer variables
     std::array<double, 3> CoMPositionInterface;
@@ -107,7 +107,7 @@ void HumanStateWrapper::run()
     pImpl->baseName = pImpl->iHumanState->getBaseName();
 
     // Prepare the message
-    human::HumanState& humanStateData = pImpl->outputPort.prepare();
+    hde::msgs::HumanState& humanStateData = pImpl->outputPort.prepare();
 
     // Convert the COM position
     humanStateData.CoMPositionWRTGlobal = {
